+++ server/util.go
@@ -24,7 +24,6 @@ import (
 	dbm "github.com/cometbft/cometbft-db"
 	tmcmd "github.com/cometbft/cometbft/cmd/cometbft/commands"
 	tmcfg "github.com/cometbft/cometbft/config"
-	tmcli "github.com/cometbft/cometbft/libs/cli"
 	tmlog "github.com/cometbft/cometbft/libs/log"
 	tmtypes "github.com/cometbft/cometbft/types"
 
@@ -174,14 +173,13 @@ func InterceptConfigsPreRunHandler(cmd *cobra.Command, customAppConfigTemplate s
 			}
 
 			opts = append(opts, log.FilterOption(filterFunc))
-		case serverCtx.Viper.GetBool(tmcli.TraceFlag):
-			// Check if the CometBFT flag for trace logging is set if it is then setup a tracing logger in this app as well.
-			// Note it overrides log level passed in `log_levels`.
-			opts = append(opts, log.LevelOption(zerolog.TraceLevel))
 		default:
 			opts = append(opts, log.LevelOption(logLvl))
 		}
 	}
+
+	// Check if the CometBFT flag for trace logging is set and enable stack traces if so.
+	opts = append(opts, log.TraceOption(serverCtx.Viper.GetBool("trace"))) // cmtcli.TraceFlag
 
 	logger := log.NewLogger(tmlog.NewSyncWriter(os.Stdout), opts...).With(log.ModuleKey, "server")
 	serverCtx.Logger = serverlog.CometLoggerWrapper{Logger: logger}